*: add --no-required option
Terraform implies a variable being required if you don't specify a default value, see [1].
Only if you leave out a default value the above "Required" column becomes a meaning.

The reality is, we specify a lot of default values for variables that are indeed required
but for whom we provide a sane default.

Furthermore we also specify quite some default empty string values "" for variables
which indeed are required, but we still cannot leave them out to mark them as required,
because of limitations in terraform where those values are being used in ternary operators.

Therefore the "Required" markdown column is meaningless in a lot of use
cases. This fixes it by introducing a `--no-required` option for markdown
generation which simply does not print it. It defaults to
false to retain current functionality.

[1] https://www.terraform.io/intro/getting-started/variables.html#defining-variables