calculate absolute path of modules
We need the absolute path of the object we are working on to set the source of the module correctly.
Sadly it will be an absolute path but either that or force users to call terrafom-docs in the dir where the module's source is set as it will be (most likely) relative eg. `source = "../../modules/bar"`