Rename 'input' to Terraform known term 'variable'
BREAKING CHANGE: - We've decided to rename `input` to `variable` which
is a known terminology in Terraform configuration. The affected parts
are: 1) markdown titles have been changed to `## Variables` 2) json item
has been renamed to `variables`.
- Also flag `--sort-inputs-by-required` has been deprecated in favor
of `--sort-by-required`. The depracted flag will be completely
removed in the second release from now.

Originally-Authored-By: Thomas Alton <thomas.alton@gmail.com>
Original-Pull-Request: #113