Move internal/tfconfig out of tree
Originally we in-tree forked hashicorp/terraform-config-inspect project
due to it being stalled and we depended on a fix upstream. But now that
seems to be resolved and they've added the feature we were looking for
in upstream.

Although, to prevent future dependency lock, we went ahead and forked
the project under terraform-docs organization and will be using that,
which at this point is identical to upstream and we don have any reason
to believe that changes anytime soon too.

Signed-off-by: Khosrow Moossavi <khos2ow@gmail.com>