Update download instruction to use tar.gz
We recently started using GoReleaser for release workflow and to be able
to support all the use-cases (brew, scopp, etc) the preferred way of
assets distribution is gonna be a packaged format (tar.z and zip).

For the time being the standalone binary will also be included in GitHub
release assets, but will be deprecated in the upcoming releases.

Signed-off-by: Khosrow Moossavi <khos2ow@gmail.com>