Use 'latest' as resource link if not a fixed version
If version contraints of required providers contain anything except of
fixed version (e.g. '1.2.3', or '= 1.2.3'), there's no way to extract
the actual intended version while being offline and solely by looking at
the terraform files, as such if such a version was provided we're going
to fall back to 'latest' as part of resource URL.

The link, unfortunately, will not be 100% accurate all the times. For
example constraint might be '>= 1.2.3, < 2.0.0' but the latest version
of said resource is at 3.x. But the silver lining is that the generated
link is never broken.

Signed-off-by: Khosrow Moossavi <khos2ow@gmail.com>