Make module source urls clickable in Markdown output
The main reason this was not already implemented, as outlined in #504 and
linked discussions, was that modules sourced from git repositories, with
urls like `git@github.com:...`, was rendered by Markdown as URL to an
email address, if the url was just shown as `[.source](.source)`.

The approach here is to detect if the module source starts with "git@",
and, if so, replace that prefix with "https://", as well as replacing the
first ":" with a "/".

Signed-off-by: Tomas Aschan <tomasl@spotify.com>