Upgrade Google Cloud Go packages in go.sum
Numerous Google Cloud Go packages have been upgraded to their latest versions in the go.sum dependencies file. This enhances the codebase with the latest features, improvements, and bug fixes offered by these packages. It's part of an effort to keep the project's dependencies up to date and ensure the application runs optimally with the latest available resources.

Signed-off-by: Aurelian Shuttleworth <aurelian@shuttleworth.tech>