Update OpenSSL package to mitigate CVE-2023-5363
OpenSSL package was updated in the Dockerfile to mitigate potential threats associated with vulnerability CVE-2023-5363. Specifically, the command "RUN apk add --no-cache --upgrade "openssl>=3.1.4-r1"" was added to ensure an updated version of OpenSSL is used which addresses the mentioned vulnerability. This enhances the overall security of our Docker containerised application.

Signed-off-by: Aurelian Shuttleworth <aurelian@shuttleworth.tech>